web/requirements/local.txt
-r base.txt

django-debug-toolbar==3.7.0
coverage==6.5.0

flake8
flake8-builtins
flake8-bugbear
pep8-naming
flake8-commas
flake8-variables-names

mypy
django-stubs
djangorestframework-stubs